requirements.txt
snowflake-connector-python
pandas
boto3
requests
PyYAML